// src/api/generated/models/category-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Ankieter API
 * Ankieter 6.0 - the API contract, source of truth for the backend and the frontend
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CategoryDto
 */
export interface CategoryDto {
    /**
     * 
     * @type {string}
     * @memberof CategoryDto
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof CategoryDto
     */
    'codePrefix': string;
    /**
     * 
     * @type {string}
     * @memberof CategoryDto
     */
    'name': string;
    /**
     * 
     * @type {number}
     * @memberof CategoryDto
     */
    'displayOrder': number;
    /**
     * 
     * @type {boolean}
     * @memberof CategoryDto
     */
    'active': boolean;
    /**
     * How many questions use it. The interface needs this to say what deactivating would affect, and to know when the prefix is still editable.
     * @type {number}
     * @memberof CategoryDto
     */
    'questionCount': number;
}